# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Tractor Work Tracker
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workers:
    get:
      operationId: listWorkers
      tags: [workers]
      summary: List all workers with payment summary
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of workers with summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkerSummary"
    post:
      operationId: createWorker
      tags: [workers]
      summary: Create a new worker
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkerBody"
      responses:
        "201":
          description: Worker created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Mobile number already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerConflictResponse"
  /workers/{id}:
    get:
      operationId: getWorker
      tags: [workers]
      summary: Get single worker details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Worker detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerSummary"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workers/{id}/logs:
    get:
      operationId: getWorkerLogs
      tags: [workers]
      summary: Get work logs for a worker
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Work logs for worker
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkLog"
  /workers/{id}/payments:
    get:
      operationId: getWorkerPayments
      tags: [workers]
      summary: Get payment history for a worker
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Worker:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
    WorkerSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: ["string", "null"]
        createdAt:
          type: string
        totalEarned:
          type: number
        totalPaid:
          type: number
        pendingAmount:
          type: number
      required:
        - id
        - name
        - createdAt
        - totalEarned
        - totalPaid
        - pendingAmount
    CreateWorkerBody:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
      required:
        - name
        - mobile
    WorkerConflictResponse:
      type: object
      properties:
        error:
          type: string
        existingWorker:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            mobile:
              type: string
          required:
            - id
            - name
      required:
        - error
        - existingWorker
    WorkLog:
      type: object
      properties:
        id:
          type: integer
        workerId:
          type: integer
        workerName:
          type: ["string", "null"]
        fieldName:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        totalHours:
          type: number
        amount:
          type: number
        paidAmount:
          type: number
        status:
          type: string
          enum: [PENDING, PARTIAL, PAID]
        jcbUserId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - workerId
        - fieldName
        - startTime
        - endTime
        - totalHours
        - amount
        - paidAmount
        - status
        - createdAt
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        workerId:
          type: integer
        amountPaid:
          type: number
        paymentDate:
          type: string
      required:
        - id
        - workerId
        - amountPaid
        - paymentDate
...