# lib/api-zod/src/generated/types/workerConflictResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Tractor Work Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { WorkerConflictResponseExistingWorker } from "./workerConflictResponseExistingWorker";

export interface WorkerConflictResponse {
  error: string;
  existingWorker: WorkerConflictResponseExistingWorker;
}